# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CourseDifficulty =
  (typeof CourseDifficulty)[keyof typeof CourseDifficulty];

export const CourseDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Course {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: CourseDifficulty;
  lessonCount: number;
  totalXp: number;
  iconColor: string;
  createdAt: string;
}

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  description: string;
  order: number;
  xpReward: number;
  challengeCount: number;
}

export interface CourseDetail {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  lessonCount: number;
  totalXp: number;
  iconColor: string;
  createdAt: string;
  lessons: Lesson[];
}

export type ChallengeType = (typeof ChallengeType)[keyof typeof ChallengeType];

export const ChallengeType = {
  multiple_choice: "multiple_choice",
  numeric_input: "numeric_input",
  true_false: "true_false",
} as const;

export interface ChallengeOption {
  id: string;
  text: string;
  isCorrect: boolean;
}

export interface Challenge {
  id: number;
  lessonId: number;
  order: number;
  type: ChallengeType;
  question: string;
  explanation: string;
  xpReward: number;
  /** @nullable */
  options?: ChallengeOption[] | null;
  correctAnswer: string;
  /** @nullable */
  hint?: string | null;
}

export interface LessonDetail {
  id: number;
  courseId: number;
  title: string;
  description: string;
  order: number;
  xpReward: number;
  challenges: Challenge[];
}

export interface CompleteChallengeBody {
  challengeId: number;
  lessonId: number;
  answer: string;
  attemptNumber: number;
}

export interface ChallengeResult {
  correct: boolean;
  xpEarned: number;
  totalXp: number;
  explanation: string;
  /** @nullable */
  hint?: string | null;
  streakDays: number;
}

export interface UserProgress {
  userId: string;
  totalXp: number;
  streakDays: number;
  lessonsCompleted: number;
  challengesCompleted: number;
  /** @nullable */
  lastActiveDate?: string | null;
}

export interface LessonProgress {
  lessonId: number;
  completed: boolean;
  completedChallenges: number[];
  xpEarned: number;
}

export interface CourseProgress {
  courseId: number;
  completedLessons: number;
  totalLessons: number;
  xpEarned: number;
  percentComplete: number;
  /** @nullable */
  currentLessonId?: number | null;
}

export interface UserProfile {
  id: string;
  clerkId: string;
  displayName: string;
  totalXp: number;
  streakDays: number;
  joinedAt: string;
}

export interface DayActivity {
  day: string;
  active: boolean;
  xpEarned: number;
}

export interface DashboardSummary {
  totalXp: number;
  streakDays: number;
  lessonsCompleted: number;
  challengesCompleted: number;
  weeklyActivity: DayActivity[];
  recommendedCourse?: Course;
  recentCourses: CourseProgress[];
}