# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all courses
 */
export const ListCoursesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  lessonCount: zod.number(),
  totalXp: zod.number(),
  iconColor: zod.string(),
  createdAt: zod.string(),
});
export const ListCoursesResponse = zod.array(ListCoursesResponseItem);

/**
 * @summary Get a course with its lessons
 */
export const GetCourseParams = zod.object({
  courseId: zod.coerce.number(),
});

export const GetCourseResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
  difficulty: zod.string(),
  lessonCount: zod.number(),
  totalXp: zod.number(),
  iconColor: zod.string(),
  createdAt: zod.string(),
  lessons: zod.array(
    zod.object({
      id: zod.number(),
      courseId: zod.number(),
      title: zod.string(),
      description: zod.string(),
      order: zod.number(),
      xpReward: zod.number(),
      challengeCount: zod.number(),
    }),
  ),
});

/**
 * @summary List lessons for a course
 */
export const ListLessonsParams = zod.object({
  courseId: zod.coerce.number(),
});

export const ListLessonsResponseItem = zod.object({
  id: zod.number(),
  courseId: zod.number(),
  title: zod.string(),
  description: zod.string(),
  order: zod.number(),
  xpReward: zod.number(),
  challengeCount: zod.number(),
});
export const ListLessonsResponse = zod.array(ListLessonsResponseItem);

/**
 * @summary Get a lesson with its challenges
 */
export const GetLessonParams = zod.object({
  lessonId: zod.coerce.number(),
});

export const GetLessonResponse = zod.object({
  id: zod.number(),
  courseId: zod.number(),
  title: zod.string(),
  description: zod.string(),
  order: zod.number(),
  xpReward: zod.number(),
  challenges: zod.array(
    zod.object({
      id: zod.number(),
      lessonId: zod.number(),
      order: zod.number(),
      type: zod.enum(["multiple_choice", "numeric_input", "true_false"]),
      question: zod.string(),
      explanation: zod.string(),
      xpReward: zod.number(),
      options: zod
        .array(
          zod.object({
            id: zod.string(),
            text: zod.string(),
            isCorrect: zod.boolean(),
          }),
        )
        .nullish(),
      correctAnswer: zod.string(),
      hint: zod.string().nullish(),
    }),
  ),
});

/**
 * @summary Get current user progress summary
 */
export const GetMyProgressResponse = zod.object({
  userId: zod.string(),
  totalXp: zod.number(),
  streakDays: zod.number(),
  lessonsCompleted: zod.number(),
  challengesCompleted: zod.number(),
  lastActiveDate: zod.string().nullish(),
});

/**
 * @summary Mark a challenge as completed and award XP
 */
export const CompleteChallengeBody = zod.object({
  challengeId: zod.number(),
  lessonId: zod.number(),
  answer: zod.string(),
  attemptNumber: zod.number(),
});

export const CompleteChallengeResponse = zod.object({
  correct: zod.boolean(),
  xpEarned: zod.number(),
  totalXp: zod.number(),
  explanation: zod.string(),
  hint: zod.string().nullish(),
  streakDays: zod.number(),
});

/**
 * @summary Get user's progress within a specific lesson
 */
export const GetLessonProgressParams = zod.object({
  lessonId: zod.coerce.number(),
});

export const GetLessonProgressResponse = zod.object({
  lessonId: zod.number(),
  completed: zod.boolean(),
  completedChallenges: zod.array(zod.number()),
  xpEarned: zod.number(),
});

/**
 * @summary Get user's progress in a course
 */
export const GetCourseProgressParams = zod.object({
  courseId: zod.coerce.number(),
});

export const GetCourseProgressResponse = zod.object({
  courseId: zod.number(),
  completedLessons: zod.number(),
  totalLessons: zod.number(),
  xpEarned: zod.number(),
  percentComplete: zod.number(),
  currentLessonId: zod.number().nullish(),
});

/**
 * @summary Get or create user profile
 */
export const GetUserProfileResponse = zod.object({
  id: zod.string(),
  clerkId: zod.string(),
  displayName: zod.string(),
  totalXp: zod.number(),
  streakDays: zod.number(),
  joinedAt: zod.string(),
});

/**
 * @summary Get dashboard summary with streak, XP, and recommended course
 */
export const GetDashboardSummaryResponse = zod.object({
  totalXp: zod.number(),
  streakDays: zod.number(),
  lessonsCompleted: zod.number(),
  challengesCompleted: zod.number(),
  weeklyActivity: zod.array(
    zod.object({
      day: zod.string(),
      active: zod.boolean(),
      xpEarned: zod.number(),
    }),
  ),
  recommendedCourse: zod
    .object({
      id: zod.number(),
      title: zod.string(),
      description: zod.string(),
      category: zod.string(),
      difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
      lessonCount: zod.number(),
      totalXp: zod.number(),
      iconColor: zod.string(),
      createdAt: zod.string(),
    })
    .optional(),
  recentCourses: zod.array(
    zod.object({
      courseId: zod.number(),
      completedLessons: zod.number(),
      totalLessons: zod.number(),
      xpEarned: zod.number(),
      percentComplete: zod.number(),
      currentLessonId: zod.number().nullish(),
    }),
  ),
});