# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeOption } from "./challengeOption";
import type { ChallengeType } from "./challengeType";

export interface Challenge {
  id: number;
  lessonId: number;
  order: number;
  type: ChallengeType;
  question: string;
  explanation: string;
  xpReward: number;
  /** @nullable */
  options?: ChallengeOption[] | null;
  correctAnswer: string;
  /** @nullable */
  hint?: string | null;
}