# lib/api-zod/src/generated/types/challengeOption.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeOption {
  id: string;
  text: string;
  isCorrect: boolean;
}