# lib/api-zod/src/generated/types/challengeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeResult {
  correct: boolean;
  xpEarned: number;
  totalXp: number;
  explanation: string;
  /** @nullable */
  hint?: string | null;
  streakDays: number;
}