# lib/api-zod/src/generated/types/completeChallengeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CompleteChallengeBody {
  challengeId: number;
  lessonId: number;
  answer: string;
  attemptNumber: number;
}