# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CourseDifficulty } from "./courseDifficulty";

export interface Course {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: CourseDifficulty;
  lessonCount: number;
  totalXp: number;
  iconColor: string;
  createdAt: string;
}