# lib/api-zod/src/generated/types/courseDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from "./lesson";

export interface CourseDetail {
  id: number;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  lessonCount: number;
  totalXp: number;
  iconColor: string;
  createdAt: string;
  lessons: Lesson[];
}