# lib/api-zod/src/generated/types/courseProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CourseProgress {
  courseId: number;
  completedLessons: number;
  totalLessons: number;
  xpEarned: number;
  percentComplete: number;
  /** @nullable */
  currentLessonId?: number | null;
}