# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Course } from "./course";
import type { CourseProgress } from "./courseProgress";
import type { DayActivity } from "./dayActivity";

export interface DashboardSummary {
  totalXp: number;
  streakDays: number;
  lessonsCompleted: number;
  challengesCompleted: number;
  weeklyActivity: DayActivity[];
  recommendedCourse?: Course;
  recentCourses: CourseProgress[];
}