# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  description: string;
  order: number;
  xpReward: number;
  challengeCount: number;
}